test(libeval): split tee-writer by behaviour family
Split tee-writer.test.js (559 LOC) into tee-writer-raw.test.js
(construction + raw mode) and tee-writer-supervised.test.js
(supervised mode) along the mode behaviour seam. Assertions unchanged;
14 tests preserved.

Spec 0640 § C (large files), design Decision 5 (≤400 LOC, split by
behaviour family).
