+++ node_modules
@@ -1,0 +1,1 @@
+/home/user/monorepo/node_modules
